requirements.txt
# Web Application Framework
flask==2.0.1            # Web framework for building the application
Werkzeug==2.0.1         # WSGI utility library used by Flask

# Resume Processing
PyPDF2==3.0.1           # Library for handling PDF files (resume parsing)

# Zoom API Integration  
requests==2.26.0        # Makes HTTP requests to external APIs (Zoom)
PyJWT==2.8.0            # JSON Web Token implementation for auth

# Configuration & System
python-dotenv==0.19.0   # Loads environment variables from .env files
zoneinfo;python_version<"3.9"  # Timezone handling for Python

# Email and Time Handling
email-validator==1.1.3  # Email validation